\name{LifeTables-internal}

\alias{get.available.indicators}
\alias{guiToolkit.default}
\alias{generateLT}
\alias{get.life.table}
\alias{updateLT}
\alias{exportLT}
\alias{plotLTcolumn}
\alias{create.plot.window}
\alias{saveGraph}
\alias{create.help.button}
\alias{plot.LifeTable}

\alias{averages.smooth}
\alias{class5}
\alias{f.csd.weight}
\alias{Mx.svd.scores}
\alias{flt.mx}
\alias{flt.mx.info}
\alias{hi.devs}
\alias{hmd.1m0.train.f}
\alias{hmd.1m0.train.m}
\alias{hmd.1m0a.train.f}
\alias{hmd.1m0a.train.m}
\alias{hmd.5m0.train.f}
\alias{hmd.5m0.train.m}
\alias{hmd.5m0a.train.f}
\alias{hmd.5m0a.train.m}
\alias{hmd.1q0.train.f}
\alias{hmd.1q0.train.m}
\alias{hmd.1q0a.train.f}
\alias{hmd.1q0a.train.m}
\alias{hmd.5q0.train.f}
\alias{hmd.5q0.train.m}
\alias{hmd.5q0a.train.f}
\alias{hmd.5q0a.train.m}
\alias{lo.devs}
\alias{mlt.mx}
\alias{mlt.mx.info}
\alias{opt.alpha.f}
\alias{opt.alpha.m}
\alias{Patterns.final.coeffs}

\title{
Internal Functions and data sets of LifeTables
}

\description{
These functions and data sets are not to be used directly by the user.
}

\keyword{internal}